# src/server/docs/swagger.yaml
openapi: 3.0.0
info:
  title: RuneRogue Game Server API
  version: 1.0.0
  description: API documentation for the RuneRogue TypeScript game server, Flask Legacy Backend, and Economy API.
servers:
  - url: http://localhost:3001
    description: Development server
paths:
...
  /players:
    get:
      summary: Get a list of players
      description: Retrieves a list of all players from the Economy API.
      tags:
        - Economy API
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
          description: Number of records to skip.
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
          description: Maximum number of records to return.
      responses:
        '200':
          description: A list of players.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayerResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create a new player
      description: Creates a new player in the Economy API.
      tags:
        - Economy API
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                is_active:
                  type: boolean
                  default: true
              required:
                - username
                - email
      responses:
        '200':
          description: Player created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Conflict, player with this email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /players/{player_id}:
    get:
      summary: Get player by ID
      description: Retrieves a single player by their unique ID from the Economy API.
      tags:
        - Economy API
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: player_id
          schema:
            type: integer
          required: true
          description: Unique ID of the player.
      responses:
        '200':
          description: Player details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Player not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /items:
    get:
      summary: Get a list of items
      description: Retrieves a list of all available items in the game from the Economy API.
      tags:
        - Economy API
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
          description: Number of records to skip.
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
          description: Maximum number of records to return.
        - in: query
          name: is_tradeable
          schema:
            type: boolean
            nullable: true
          description: Filter by tradeable status.
        - in: query
          name: name_contains
          schema:
            type: string
            nullable: true
          description: Filter by item name containing this string.
      responses:
        '200':
          description: A list of items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create a new item
      description: Creates a new item in the Economy API.
      tags:
        - Economy API
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                  nullable: true
                is_tradeable:
                  type: boolean
                  default: true
                is_stackable:
                  type: boolean
                  default: false
                base_value:
                  type: number
                  format: float
                  default: 0
              required:
                - name
      responses:
        '200':
          description: Item created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Conflict, item with this name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /items/{item_id}:
    get:
      summary: Get item by ID
      description: Retrieves a single item by its unique ID from the Economy API.
      tags:
        - Economy API
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: item_id
          schema:
            type: integer
          required: true
          description: Unique ID of the item.
      responses:
        '200':
          description: Item details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Item not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /players/{player_id}/inventory:
    get:
      summary: Get player's inventory
      description: Retrieves a player's inventory from the Economy API.
      tags:
        - Economy API
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: player_id
          schema:
            type: integer
          required: true
          description: Unique ID of the player.
      responses:
        '200':
          description: A list of inventory items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItemResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Player not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Add item to player's inventory
      description: Adds an item to a player's inventory in the Economy API.
      tags:
        - Economy API
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: player_id
          schema:
            type: integer
          required: true
          description: Unique ID of the player.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                player_id:
                  type: integer
                item_id:
                  type: integer
                quantity:
                  type: integer
                  default: 1
              required:
                - player_id
                - item_id
      responses:
        '200':
          description: Item added to inventory successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItemResponse'
        '400':
          description: Bad Request, e.g., player ID mismatch or non-stackable item already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Player or Item not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /trades:
    get:
      summary: Get a list of trades
      description: Retrieves a list of trades from the Economy API, with optional filtering.
      tags:
        - Economy API
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            nullable: true
            enum: [pending, accepted, rejected, completed, cancelled]
          description: Filter trades by status.
        - in: query
          name: player_id
          schema:
            type: integer
            nullable: true
          description: Filter trades by initiator or receiver player ID.
      responses:
        '200':
          description: A list of trades.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    initiator_id:
                      type: integer
                    receiver_id:
                      type: integer
                    status:
                      type: string
                    initiated_at:
                      type: string
                      format: date-time
                    completed_at:
                      type: string
                      format: date-time
                      nullable: true
                    items:
                      type: array
                      items:
                        type: object
                        properties:
                          item_id:
                            type: integer
                          quantity:
                            type: integer
                          from_player_id:
                            type: integer
                          to_player_id:
                            type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Not authenticated
  schemas:
    PlayerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the player.
        username:
          type: string
          description: Player's username.
        email:
          type: string
          format: email
          description: Player's email address.
        is_active:
          type: boolean
          description: Whether the player account is active.
        created_at:
          type: string
          format: date-time
          description: Timestamp when the player account was created.
        last_login:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the player's last login.
      required:
        - id
        - username
        - email
        - is_active
        - created_at
    ItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the item.
        name:
          type: string
          description: Name of the item.
        description:
          type: string
          nullable: true
          description: Description of the item.
        is_tradeable:
          type: boolean
          description: Whether the item can be traded.
        is_stackable:
          type: boolean
          description: Whether the item can be stacked in inventory.
        base_value:
          type: number
          format: float
          description: Base value of the item.
        created_at:
          type: string
          format: date-time
          description: Timestamp when the item was created.
      required:
        - id
        - name
        - is_tradeable
        - is_stackable
        - base_value
        - created_at
    InventoryItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the inventory item entry.
        player_id:
          type: integer
          format: int64
          description: ID of the player who owns the item.
        item_id:
          type: integer
          format: int64
          description: ID of the item.
        quantity:
          type: integer
          description: Quantity of the item.
        acquired_at:
          type: string
          format: date-time
          description: Timestamp when the item was acquired.
      required:
        - id
        - player_id
        - item_id
        - quantity
        - acquired_at
    Error:
      type: object
      properties:
        error:
          type: string
          description: A descriptive error message.
          example: "An unexpected error occurred."
        details:
          type: string
          description: Optional. More detailed information about the error.
...